A top level view can return null
